+++ asteroid-core/src/main/java/asteroid/transformer/AbstractExpressionTransformer.java
@@ -2,8 +2,10 @@ package asteroid.transformer;
 
 import groovy.lang.Closure;
 
+import org.codehaus.groovy.syntax.Types;
 import org.codehaus.groovy.control.SourceUnit;
 import org.codehaus.groovy.ast.expr.Expression;
+import org.codehaus.groovy.ast.expr.BinaryExpression;
 import org.codehaus.groovy.ast.expr.MethodCallExpression;
 
 /**
@@ -20,7 +22,8 @@ import org.codehaus.groovy.ast.expr.MethodCallExpression;
  */
 public abstract class AbstractExpressionTransformer<T extends Expression> extends AbstractTransformer {
 
-    private final Closure<Boolean> criteria;;
+    private final Closure<Boolean> criteria;
+    private final Class<T> clazz;
 
     /**
      * Every instance needs the source unit awareness and the name of the method
@@ -31,9 +34,42 @@ public abstract class AbstractExpressionTransformer<T extends Expression> extend
      * {@link Expression}
      * @since 0.2.0
      */
+    @Deprecated
     public AbstractExpressionTransformer(final SourceUnit sourceUnit, final Closure<Boolean> criteria) {
         super(sourceUnit);
         this.criteria = criteria;
+        this.clazz = null;
+    }
+
+    /**
+     * Every instance needs the source unit awareness and the name of the method
+     * it's going to transform
+     *
+     * @param clazz the type of the expression we're interested in
+     * @param sourceUnit Needed to apply scope
+     * @param criteria the criteria used to search the interesting
+     * {@link Expression}
+     * @since 0.2.3
+     */
+    public AbstractExpressionTransformer(final Class<T> clazz, final SourceUnit sourceUnit, final Closure<Boolean> criteria) {
+        super(sourceUnit);
+        this.clazz = clazz;
+        this.criteria = criteria;
+    }
+
+    /**
+     * Every instance needs the source unit awareness and the name of the method
+     * it's going to transform
+     *
+     * @param clazz the type of the expression we're interested in
+     * @param sourceUnit Needed to apply scope
+     * {@link Expression}
+     * @since 0.2.3
+     */
+    public AbstractExpressionTransformer(final Class<T> clazz, final SourceUnit sourceUnit) {
+        super(sourceUnit);
+        this.clazz = clazz;
+        this.criteria = everything();
     }
 
     /**
@@ -45,9 +81,21 @@ public abstract class AbstractExpressionTransformer<T extends Expression> extend
             return null;
         }
 
-        final boolean proceed = criteria.call(expression);
+        final boolean matches = isOfTypeAndMatchesCriteria(expression, clazz, criteria);
 
-        return proceed ? this.transformExpression((T) expression) : expression.transformExpression(this);
+        if (matches) {
+            return this.transformExpression((T) expression);
+        }
+
+        return expression.transformExpression(this);
+    }
+
+    private static boolean isOfTypeAndMatchesCriteria(final Expression expression, final Class clazz, final Closure<Boolean> predicate) {
+        return isOfType(expression, clazz) && predicate.call(expression);
+    }
+
+    private static boolean isOfType(final Expression expression, final Class clazz) {
+        return clazz != null && clazz.isInstance(expression);
     }
 
     /**
@@ -91,6 +139,32 @@ public abstract class AbstractExpressionTransformer<T extends Expression> extend
         return new Closure<Boolean>(null) {
             public Boolean doCall(final Expression expression) {
                 return true;
+            }
+        };
+    }
+
+    /**
+     * Checks that a given {@link BinaryExpression} uses a specific
+     * token type. The token type is an `int` value. You can use
+     * {@link Types} where all token types are declared.
+     *
+     * @param tokenType Check {@link Types} for more info
+     * @return a {@link Closure} used as criteria
+     * @since 0.2.3
+     * @see Types
+     */
+    public static Closure<Boolean> usesToken(final int tokenType) {
+        return new Closure<Boolean>(null) {
+            public Boolean doCall(final Expression expression) {
+                if (!(expression instanceof BinaryExpression)) {
+                    return false;
+                }
+
+                final BinaryExpression binaryExpression = (BinaryExpression) expression;
+
+                return binaryExpression
+                    .getOperation()
+                    .getType() == tokenType;
             }
         };
     }