+++ asteroid-test/src/main/groovy/asteroid/global/samples/ChangeTripleXToPlusOne.groovy
@@ -15,7 +15,7 @@ class ChangeTripleXToPlusOne
     extends AbstractExpressionTransformer<MethodCallExpression> { // <1>
 
     ChangeTripleXToPlusOne(final SourceUnit sourceUnit) {
-        super(sourceUnit, methodCallByNameEq('xxx')) // <2>
+        super(MethodCallExpression, sourceUnit, methodCallByNameEq('xxx')) // <2>
     }
 
     @Override