+++ asteroid-test/src/main/groovy/asteroid/statements/trycatch/Log.groovy
@@ -1,0 +1,8 @@
+package asteroid.statements.trycatch
+
+interface Log {
+
+    void info(Object o)
+
+    void error(Object o)
+}