+++ asteroid-test/src/main/groovy/asteroid/statements/trycatch/Safe.groovy
@@ -1,0 +1,8 @@
+package asteroid.statements.trycatch
+
+import asteroid.Local
+
+@Local(value = SafeImpl, applyTo = Local.TO.METHOD)
+@interface Safe {
+    int value()
+}