+++ asteroid-test/src/test/groovy/asteroid/statements/trycatch/TryCatchStatementSpecExample.groovy
@@ -1,0 +1,9 @@
+package asteroid.statements.trycatch
+
+class TryCatchStatementSpecExample {
+
+    @Safe(100)
+    Integer add(Integer one, Integer two) {
+        return one + two
+    }
+}